+++ libraries/effect/src/main/java/androidx/media3/effect/GaussianBlurWithFrameOverlaid.java
@@ -17,7 +17,6 @@ package androidx.media3.effect;
 
 import android.content.Context;
 import androidx.annotation.FloatRange;
-import androidx.annotation.RequiresApi;
 import androidx.media3.common.VideoFrameProcessingException;
 import androidx.media3.common.util.UnstableApi;
 
@@ -27,7 +26,6 @@ import androidx.media3.common.util.UnstableApi;
  * <p>The width of the blur is specified in pixels and applied symmetrically.
  */
 @UnstableApi
-@RequiresApi(26) // See SeparableConvolution.
 public final class GaussianBlurWithFrameOverlaid extends SeparableConvolution {
   private final float sigma;
   private final float numStandardDeviations;